Get the series from CBRT

Get the series from CBRT

cbRt_get(series, startDate, endDate, token, aggregationTypes = NULL,
  formulas = NULL, freq = NULL, nd = TRUE, as = c("tibble",
  "tibbletime", "data.frame", "data.table"))

Arguments

series
Series Code series argument can be obtained from CBRT webpage. Search CBRT webpage in order to find out the series code and use it in the series argument. The serial codes are displayed. If more than one series is selected, it takes up parameters as the number of series, the serial codes are separated by a "-" sign.

startDate
Start date startDate argument is the series start date as dd-mm-yyyy format. In order to display the frequency of the desired series, the first day of the corresponding frequency must be stated in the start date as dd-mm-yyyy format.

endDate
End date endDate argument is the series end date as dd-mm-yyyy format.

token
API key token argument is the required API key. See <https://evds2.tcmb.gov.tr/help/videos/EVDS_Web_Service_Usage_Guide.pdf> for instructions to obtain the API key.

aggregationTypes
Aggregation type aggregationTypes argument is the aggregation applied to series. Available aggregations are; avg : Average min : Minimum max : Maximum first : Beginning last : End sum : Cumulative If more than one series is selected, it takes up parameters as the number of series, the serial codes are separated by a "-" sign. If this parameter is not entered by the user, the original observation parameter is applied for the relevant series.

formulas
Formulas applied to series formulas argument is the formula applied to series. Available formulas are; 0: Level 1: Percentage change 2: Difference 3: Year-to-year Percent Change 4: Year-to-year Differences 5: Percentage Change Compared to End-of-Previous Year 6: Difference Compared to End-of-Previous Year 7: Moving Average 8: Moving Sum If more than one series is selected, it takes up parameters as the number of series, the serial codes are separated by a "-" sign. If this parameter is not entered by the user, the level formula parameter is applied for the relevant series.

freq
Frequency of series freq argument is the frequency of the series. Available frequencies are; 1: Daily 2: Business 3: Weekly (Friday) 4: Twicemonthly 5: Monthly 6: Quarterly 7: Semiannual 8: Annual This parameter takes a single value. If this parameter is not entered by the user, the common frequency of the series is taken. If you enter a higher frequency (eg: monthly) than the common frequency of the series (eg: annually), the common frequency of the series is taken into account (eg: annually).

nd
Convert ND values to NA nd argument is a TRUE or FALSE argument. Data retrieved sometimes includes ND terms. If nd is set to TRUE, all NDs are converted to NAs.

Value

Argument as can be set to either tibble, tibbletime, data.frame, or data.table to obtain different types of data classes. tibble is the default output class.

Examples

# Download the given series for the given dates

# NOT RUN {
get_series(series = "TP.DK.USD.A",
                    startDate = "01-01-2017",
                    endDate = "01-01-2018",
                    token = APIkey)
# }